packages/ai-service/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
pinecone-client==3.0.0
sentence-transformers==2.2.2
pydantic==2.5.3
redis==5.0.1
psycopg2-binary==2.9.9
pdf2image==1.16.3
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
